TEMA - PROFESOR

+++ src/main/java/net/ausiasmarch/academia/service/TemaService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import net.ausiasmarch.academia.entity.TemaEntity;
 import net.ausiasmarch.academia.exception.ResourceNotFoundException;
+import net.ausiasmarch.academia.exception.UnauthorizedAccessException;
 import net.ausiasmarch.academia.repository.TemaRepository;
 
 @Service
@@ -22,64 +23,96 @@ public class TemaService implements ServiceInterface<TemaEntity> {
     @Autowired
     CursoService oCursoService;
 
+    @Autowired
+    AuthService oAuthService;
+
     private String[] arrTitulo = {
-        "Matemáticas Básicas",
-        "Historia Mundial",
-        "Fundamentos de Programación",
-        "Química Orgánica",
-        "Lengua y Literatura",
-        "Introducción a la Física",
-        "Análisis de Datos",
-        "Biología Celular",
-        "Psicología General",
-        "Economía Básica"
+            "Matemáticas Básicas",
+            "Historia Mundial",
+            "Fundamentos de Programación",
+            "Química Orgánica",
+            "Lengua y Literatura",
+            "Introducción a la Física",
+            "Análisis de Datos",
+            "Biología Celular",
+            "Psicología General",
+            "Economía Básica"
     };
-    
+
     private String[] arrDescripcion = {
-        "Examen que evalúa habilidades aritméticas y algebraicas fundamentales.",
-        "Prueba que abarca eventos clave y figuras históricas globales.",
-        "Evaluación de conocimientos básicos en lógica y estructuras de programación.",
-        "Examen sobre las propiedades y reacciones de compuestos orgánicos.",
-        "Prueba centrada en análisis de textos literarios y comprensión gramatical.",
-        "Evaluación de conceptos básicos de mecánica y leyes de Newton.",
-        "Examen sobre técnicas de recopilación y análisis de datos estadísticos.",
-        "Prueba que aborda estructuras y funciones celulares en organismos vivos.",
-        "Examen introductorio a teorías y conceptos en psicología.",
-        "Evaluación de principios económicos y fundamentos de oferta y demanda."
+            "Examen que evalúa habilidades aritméticas y algebraicas fundamentales.",
+            "Prueba que abarca eventos clave y figuras históricas globales.",
+            "Evaluación de conocimientos básicos en lógica y estructuras de programación.",
+            "Examen sobre las propiedades y reacciones de compuestos orgánicos.",
+            "Prueba centrada en análisis de textos literarios y comprensión gramatical.",
+            "Evaluación de conceptos básicos de mecánica y leyes de Newton.",
+            "Examen sobre técnicas de recopilación y análisis de datos estadísticos.",
+            "Prueba que aborda estructuras y funciones celulares en organismos vivos.",
+            "Examen introductorio a teorías y conceptos en psicología.",
+            "Evaluación de principios económicos y fundamentos de oferta y demanda."
     };
-    
 
     public Long randomCreate(Long cantidad) {
         for (int i = 0; i < cantidad; i++) {
             TemaEntity oTemaEntity = new TemaEntity();
-            
+
             oTemaEntity.setTitulo(arrTitulo[oRandomService.getRandomInt(0, arrTitulo.length - 1)]);
-            
+
             oTemaEntity.setDescripcion(arrDescripcion[oRandomService.getRandomInt(0, arrDescripcion.length - 1)]);
-            
+
             oTemaEntity.setCurso(oCursoService.randomSelection());
-            
+
             oTemaRepository.save(oTemaEntity);
         }
         return oTemaRepository.count();
-    } 
+    }
 
     // Cargar datos - Tema
     public Page<TemaEntity> getPage(Pageable oPageable, Optional<String> filter) {
 
-        if (filter.isPresent()) {
-            return oTemaRepository
-                    .findByTituloContainingOrDescripcionContaining(
-                            filter.get(), filter.get(),
-                            oPageable);
-        } else {
-            return oTemaRepository.findAll(oPageable);
+        if (oAuthService.isAdmin()) {
+            if (filter.isPresent()) {
+                return oTemaRepository
+                        .findByTituloContainingOrDescripcionContaining(
+                                filter.get(), filter.get(),
+                                oPageable);
+            } else {
+                return oTemaRepository.findAll(oPageable);
+            }
         }
+
+        if (oAuthService.isProfesor()) {
+            System.out.println(oAuthService.getUsuarioFromToken().getId());
+
+            if (filter.isPresent()) {
+                return oTemaRepository
+                        .findTemasByUsuarioIdAndFilter(oAuthService.getUsuarioFromToken().getId(), filter.get(),
+                                oPageable);
+            } else {
+                return oTemaRepository.findTemasByUsuarioId(oAuthService.getUsuarioFromToken().getId(), oPageable);
+            }
+        }
+
+        throw new UnauthorizedAccessException("No tienes permisos para generar la lista de temas.");
     }
 
     public TemaEntity get(Long id) {
-        return oTemaRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Tema no encontrado"));
+        if (oAuthService.isAdmin()) {
+            return oTemaRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Tema no encontrado."));
+        }
+
+        if (oAuthService.isProfesor()) {
+            if (oTemaRepository.existsTemaByUsuarioId(oAuthService.getUsuarioFromToken().getId(), id) > 0) {
+                return oTemaRepository.findById(id)
+                        .orElseThrow(() -> new ResourceNotFoundException("Tema no encontrado."));
+            } else {
+                throw new UnauthorizedAccessException("No puedes ver este tema porque no pertenece a ninguno de tus cursos.");
+            }
+        }
+
+        throw new UnauthorizedAccessException("No tienes permisos para ver este tema.");
+
     }
 
     // Cargar datos - Curso
@@ -101,31 +134,77 @@ public class TemaService implements ServiceInterface<TemaEntity> {
 
     // Crear
     public TemaEntity create(TemaEntity oTemaEntity) {
+        if (!oAuthService.isAdminOrProfesor()) {
+            throw new UnauthorizedAccessException("No tienes permisos para crear temas.");
+        }
+
         return oTemaRepository.save(oTemaEntity);
     }
 
     // Eliminar
     public Long delete(Long id) {
-        oTemaRepository.deleteById(id);
-        return 1L;
+        if (oAuthService.isAdmin()) {
+            oTemaRepository.deleteById(id);
+            return 1L;
+        }
+
+        if (oAuthService.isProfesor()) {
+
+            if (oTemaRepository.existsTemaByUsuarioId(oAuthService.getUsuarioFromToken().getId(), id) > 0) {
+                oTemaRepository.deleteById(id);
+                return 1L;
+            } else {
+                throw new UnauthorizedAccessException("No puedes eliminar este tema porque no pertenece a tus cursos.");
+            }
+        }
+
+        throw new UnauthorizedAccessException("No tienes permisos para eliminar temas.");
     }
 
     // Actualizar
     public TemaEntity update(TemaEntity oTemaEntity) {
-        TemaEntity oTemaEntityFromDatabase = oTemaRepository.findById(oTemaEntity.getId()).get();
-        if (oTemaEntity.getTitulo() != null) {
-            oTemaEntityFromDatabase.setTitulo(oTemaEntity.getTitulo());
+        if (oAuthService.isAdmin()) {
+            
+            TemaEntity oTemaEntityFromDatabase = oTemaRepository.findById(oTemaEntity.getId()).get();
+
+            if (oTemaEntity.getTitulo() != null) {
+                oTemaEntityFromDatabase.setTitulo(oTemaEntity.getTitulo());
+            }
+
+            if (oTemaEntity.getDescripcion() != null) {
+                oTemaEntityFromDatabase.setDescripcion(oTemaEntity.getDescripcion());
+            }
+
+            if (oTemaEntity.getCurso() != null) {
+                oTemaEntityFromDatabase.setCurso(oTemaEntity.getCurso());
+            }
+
+            return oTemaRepository.save(oTemaEntityFromDatabase);
         }
 
-        if (oTemaEntity.getDescripcion() != null) {
-            oTemaEntityFromDatabase.setDescripcion(oTemaEntity.getDescripcion());
+        if (oAuthService.isProfesor()) {
+            if (oTemaRepository.existsTemaByUsuarioId(oAuthService.getUsuarioFromToken().getId(),
+                    oTemaEntity.getId()) > 0) {
+
+                TemaEntity oTemaEntityFromDatabase = oTemaRepository.findById(oTemaEntity.getId()).get();
+
+                if (oTemaEntity.getTitulo() != null) {
+                    oTemaEntityFromDatabase.setTitulo(oTemaEntity.getTitulo());
+                }
+
+                if (oTemaEntity.getDescripcion() != null) {
+                    oTemaEntityFromDatabase.setDescripcion(oTemaEntity.getDescripcion());
+                }
+
+                if (oTemaEntity.getCurso() != null) {
+                    oTemaEntityFromDatabase.setCurso(oTemaEntity.getCurso());
+                }
+            } else {
+                throw new UnauthorizedAccessException("No puedes eliminar este tema porque no pertenece a tus cursos.");
+            }
         }
 
-        if (oTemaEntity.getCurso() != null) {
-            oTemaEntityFromDatabase.setCurso(oTemaEntity.getCurso());
-        }
-
-        return oTemaRepository.save(oTemaEntityFromDatabase);
+        throw new UnauthorizedAccessException("No tienes permisos para editar temas.");
     }
 
     // Random Selection
